refactor type declations
